Replace the white-space based indentation of Web IDL code with a CSS-based one
This is done to make it easier to cut and paste the Web IDL in the Web Audio
spec to use for implementing the spec.  In particular it will make things
easier when diffing the newer versions of the spec against the Web IDL files
that an implementation has been using previously.