Fix #5: Add setTargetAtTime curve to example
The graph of parameter automation is updated to include an example of
setTargetAtTime with the code update appropriately.  The graph also
includes some simple annotation of each part and also gives an example
of how to have another automation follow a setTargetAtTime.  A
setValueAtTime is required to terminate the setTargetAtTime and give a
starting value for the following exponentialRampToValueAtTime.