createIIRFilter takes double coefficients
The coefficients for createIIRFilter are explicitly doubles and not
floats now.

We also fix a minor issue that was accepted in
WebAudio/web-audio-api#323 where a sequence<float> was allowed instead
of Float32Array to make it easier for the developer to specify
coefficients.

Note that for some reason respec replaces sequence<double> with just
sequence in the document.  Is this a bug in respec?

Fix WebAudio/web-audio-api#631