Fix #769 by correcting the formulas
The formulas for the lowpass and highpass filters are wrong; they
depend on Q, not G. Update the formulas to match what browsers
actually do.

Also updated the description of Q to say that Q is in dB for lowpass
and highpass filters and linear for all others.  Because that's what's
really happening.