Fix #348: Add latencyHint
This updates and replaces the current playbackCategory to a
latencyHint that accepts an AudioContextLatencyCategory (a renaming of
AudioContextPlaybackCategory to reflect better the actual intent) or a
double for finer control of the latency.

An additional attribute, processingLatency, is added to the
BaseAudioContext to indicate what latency was actually used by the
context.