Fix #771 by using Audio Cookbook formula
Use the Audio Cookbook formula for the lowpass and highpass filters.
However, the cookbook has Q in linear units.  This change makes Q in
dB.  This preserves the old behavior which used Q in dB for the
lowpass and highpass filters, so that automations have the same value.
(See #769)

This also minimizes the difference in sound quality between these new
filters and the old ones.