Fix #821: specify behavior of setValueCurveAtTime
setValueCurveAtTime is modified so that an implicit call to
setValueAtTime is made at the end of the curve duration.  This makes
any following automation events start from the end of the
setValueCurveAtTime event with the last value of the curve.

Without this, a linearRamp following setValueCurve would replace
setValueCurve.  This is not the desired behavior.