Fix #907: Define initial up and forward vectors correctly.
Set the initial values for the up and forward vectors correctly.  This
was a mistake introduced in adding the up and forward AudioParams.

Also took the opportunity to move some of the descriptive text out of
forwardX and upX to the introduction paragraph to describe forward
and up vectors better.  They didn't really belong there anyway.  And
this makes the description of each component uniform so it's easier to
read and find the default values.

Finally, add a dfn entry for the Spatialization/Panning section so
that it's easier to refer to.  Updated links in the listener section
appropriately.