Update according to review, with some editorial changes.
Set maxDelayTime to default to 1.

The DelayOptions section should be a subsection of the DelayNode
interface.

Add a mention of the maxDelayTime member of the attribute to the
description of the delayTime attribute of the node.  This description
mentioned the parameter for the factory method, so we should a one for
the constructor dictionary.