Fix #611: Clamp maxDistance and refDistance
For the linear distance model, make d'max = max(dref, dmax) and d'ref
= min(dref,dmax).  And specify the value of the model is 1-f if
d'ref=d'max. This makes the model well-defined for all values of dref
and dmax.

Note that this change allows a user to set dref > dmax, but the
behavior is well-defined in this case (by reversing the order). Not
sure if we want that or not.

Also, specify that setting refDistance and maxDistance to a
non-positive value throws a TypeError.