Fix #324: Handle degenerate azimuth/elevation calculation
In the current code, if the source to listener vector is zero, the
algorithm immediately returns azimuth = 0 and elevation = 0.  This
causes a discontinuity.

Instead, remove the early exit and let the source-listener vector
propagate through. This removes this discontinuity, but in general
there is still the discontinuity when the source and listener
positions are the same.  This is unavoidable.

Also add a short note on what normalize() does in general and when the
vector is the 0 vector. (Does nothing).