Fix #1105: Allow refDistance to be 0 for PannerNode
If refDistance is 0, define the inverse and exponential models to take
the value 0 independent of the values of d and f.

The limit as d and dref go to zero aren't well defined and depends on
the direction of approach.  So I took the lazy way and set the value
to be always be 0.