Fix #1269: Add selectable rate for each AudioParam
For each `AudioParam` add an `automationRate` attribute that can be
set to either "a-rate" or "k-rate".  But some nodes have constraints
on the possible values so enumerate those constraints.

For each AudioParam, specify the default initial value of the
AudioParam.

Also define an "effective automation rate" for a `PannerNode` which
basically means the `AudioParam` is "k-rate" if the `panningModel` is
"HRTF", independent of the actual value of the `automationRate`.

Adjust the "listener AudioParam rate" to use the "effective autoamtion
rate" of the `PannerNode`.