Fix #1577: Add constraints in constructor methods if necessary
Just went through the factory methods that had parameters and looked
to see if there were constraints on the values of the parameters.
Then looked at the constructor options dictionary and linked to these
constraints if needed.

While we're here also fix #1588, which has the descriptions of the
members backwards.