Make it easier to maintain expected-errs
When building locally and the expected errors change, it would be
really nice if you could just copy the actual output from bikeshed to
expected-errs.txt.  Then when running travis, have travis do the
necessary things to compare the actual output with the expected.

And printing out the actual log might make it easier to find and fix
issues because bikeshed does produce line numbers for where the issues
are.

Travis takes care of removing the line numbers and such from both and
does the comparison.