Fix #1648, #1649: refine algorithm for setting buffer and curve
These algorithms for setting the buffer (ConvovlerNode) and the curve
(WaveShaperNode) should be consistent with how the buffer for an
AudioBufferSourceNode is handled.  The original intent was that you
can only set the buffer(curve) just once, but you can set them to null
as many times as you want, but you can't then set it to a non-null
buffer.