Fix #1698: Restrict nominal range of OscillatorNode.detune
The current range for the detune attribute is way too wide because the
computed oscillator frequency needs to be within the Nyquist
frequency.

The new limits allow the detune parameter to combine with any valid
frequency to result in a frequency that still covers the entire
frequency range.  Thus you can detune a frequency of -Nyquist to
Nyquist if so desired.