Fix #1788: Clarify the value returned by the AudioParam value getter.
Define the intrinsic value and update the algorithm a little bit for
that.

Then update the rendering algorith so that the [[current value]] is
the intrinsic value, not the computed value (that consists of the
automation and any inputs to the param).

I think this could be simplified a bit more, but I didn't really do
that.  Perhaps all references to intrinsic value could be replaced by
[[current value]]?