Fix #1858: Multiline source text loses space (#1940)
Use the workaround and make the paragraph all on one line.  This fixes
the issue.

I only looked through the Description column of parameter descriptions
for methods.  These are all I could find; there could be others or in
other places, but I didn't see any when proofreading the whole spec a
while ago.