Refactor the constructor for AudioNodes.
This removes the generic ctor algorithm in favour of one for each type
that calls the "initialize" algorithm, and pass the arguments
appropriately.

This catches a few mistakes, and changes the initialize algorithm itself
to be explicit about what to do in case of exceptions. The setters for
the attribute can frequently throw, so this automatically handles
invalid values by throwing the right exception. The respective order of
the operation is also clear.