Create PULL_REQUEST_TEMPLATE.md
Recent situation come up where a change was made to the spec, but no corresponding web platform test was added. This ended up causing interop issues in Gecko. To prevent this, I'll encourage the WG to adopt this template. It helps guard against changes to being made to the spec without having both a Web Platform Test and public implementation commitments from the various browser engines.