Spec document unloading
This is a rebase of the changes lost in the bikeshed conversion.

This does the following:

- Keep all promises around, so we can deal with them on unload
- Spec what to do when unloading a document: very similar to calling
  close() on all audiocontext
- This stops the rendering of AudioContext, stops decoding of
  decodeAudioData tasks, stop the rendering of OfflineAudioContext.
- A consequence of this is that there won't be any audio output anymore.
- Another consequence of this (because the rendering loop does not
  running anymore) is that any control message sent by Web Audio API
  objects are never going to be executed. We don't have to check for
  "active document" everywhere.

This fixes #1139 and #2155.

This version converts to bikeshed markup, harmonizes the names of the
slots (`[[pending promises]]` and `[[pending resume promises]]`), and
hooks the slots of the interface that makes the most sense
(respectively, `BaseAudioContext` and `AudioContext`).

There is no provision for doing anything with the `addModule` promise,
this should be in the worklet text.