Fix #2190: Update expected results (#2196)
The latest bikeshed has fixed the issues we had, and has also improved
the errors in the generating the spec.  So update expected-errs.txt
with the new results.

Also, add some command line options to compile.sh.  In particular -K
will keep actual-errs.txt so you can examine it and copy it to
expected-errs.txt as needed.  The option -h (or -?) prints out a brief
explanation of the available command-line options.