Fix #2169: Remove redundant constraints in AudioParamDescriptor (#2222)
* Fix #2169: Remove redundant constraints in AudioParamDescriptor

The constraings in the members of the `AudioParamDescriptor` are removed.  We replace that with a link to the steps in `registerProcessor` algorithm that describes how to handle the descriptor.

Updated the algorithm to include a step to throw an error if `minValue` is greater than `maxValue`.

* Remove the new step and simplify the other step

Makes it clear that min <= max.