Fix #2227: Mention ABSN.start acquires the contents of the buffer
Just make it more explicit that calling `start()` will acquire the
contents of the buffer.  Without this, you kind of have to know that
this is described in the algorithm for acquiring the contents.

Also missed a minor issue in the start algorithm to abort the steps if
any exceptions are thrown because of parameter issues, as we say for
`AudioScheduledSourceNode.start()`.