Fix #2262: Clone bikeshed and install it
As a workaround, clone the bikeshed repo and install bikeshed from the
repo.

Ideally, want just want to use the version with pip3 instead of
cloning the current version of bikeshed, but that's producing
unexpected errors.  We'll do this for now and recheck at some later
date when bikeshed is updated and revert to the old version.