Fix #2257: Make channel ordering table visible in dark mode (#2264)
When using dark mode, use a different background color for the even
rows of the channel ordering table so that the entries are visible.

While we're at it, make the headings for the audio node properties
table, the audio param values table, and the enumeration description
table look a little better in dark mode.  This is done by specifying a
brighter green to make it a bit more legible.