Fix #1933:  Use FrozenArray for AudioWorkletProcessor process() (#2250)
* Fix #1933:  Use FrozeArray for AudioWorkletProcessor process()

Instead of `sequence<sequence<Float32Array>>`, use
`FrozenArray<FrozenArray<Float32Array>>`.

Define a callback function as well.

* Address review comments from karlt.

Rename AudioWorkletProcessorCallback to AudioWorletProcessCallback

The third parameter is an object, not frozen array.

* Replace process() method with callback sections

Several changes here in no particular order:

- Tell bikeshed that `object` is the webidl concept
- Fix parameter names for the process callback (plurals, as used in the descriptions)
- Rename the dfn for "Processing an input buffer" from "process" to "processing-input-buffer".
- Fix up links to `process()` to point to the right place.
- The old section on the process method is broken up into two sections.  The main one is renaming it to "callback AudioWorkletProcessCallback".  This contains most fo the main text. A subsection then describes the parameters to the callback.

* Update expected errors

The callback idea fixes the errors from defining the process method.

* Address some review comments

* Describe what happens to inputs/outputs if things change

Add text to say inputs/outputs are recycled unless the topology
changes (number of channels, array is transferred).  Then new arrays
are allocated.

* Address some review comments

Still need description of how parameters are frozen.

* Update README.md

* Fix typo in "messsage" (#2255)

* chore: Fix markdown for bold (#2258)

* Update implementation-report.html

Update for Safari and for Edge

* Update implementation-report.html

Update IDL results for newer browsers

* Update implementation-report.html

Update summaries for new browsers. Fails = 6573 -  passes

* Update implementation-report.html

Date

* Fix typo "atleast" (#2259)

This isn't really a typo but an issue with bikeshed not leaving a
space at line breaks.  See tabatkins/bikeshed#1470.

* Address #2185: Fix duplicated IDs for decode callbacks (#2253)

* Address #2185: Duplicate IDs for Decode callbacks

The decode callbacks were multiply defining their arguments.  Instead,
just have the descriptions of the arguments just link back to the IDL
that defines the args.

* Update expected errors

* Fix #2262: Clone bikeshed and install it (#2263)

As a workaround, clone the bikeshed repo and install bikeshed from the
repo.

Ideally, want just want to use the version with pip3 instead of
cloning the current version of bikeshed, but that's producing
unexpected errors.  We'll do this for now and recheck at some later
date when bikeshed is updated and revert to the old version.

* Fix #2257: Make channel ordering table visible in dark mode (#2264)

When using dark mode, use a different background color for the even
rows of the channel ordering table so that the entries are visible.

While we're at it, make the headings for the audio node properties
table, the audio param values table, and the enumeration description
table look a little better in dark mode.  This is done by specifying a
brighter green to make it a bit more legible.

* Add ECMAScript algorithm to freeze parameters.

* Some cleanups.

Co-authored-by: Chris Lilley <chris@w3.org>
Co-authored-by: Tomohiro IKEDA <rilakkuma.san.xjapan@gmail.com>