Some cleanups
For OfflineAudioContextOptions.renderSizeHint, just say it's the same
as for AudioContextOptions, except "hardware" is the same as
"default'.

Allow more flexibility for the interaction between `latencyHint` and
`renderSize`.  I think Firefox always returns 0 for `latencyHint`
because it doesn't double-buffer.  Chrome double-buffers so the
`latencyHint` isn't zero.

Mark the example as being non-normative.