Address WebAudio/web-audio-api-v2#13: Explainer for user-selectable render size (#2351)
* Address #13: Add explainer for user-selectable size

* update markdown table for display

* Rearrange sections and add details for latencyHint

Make the API section more like the spec section.

Add a section on how renderSize interacts with latencyHint.

* More editorial stuff

* Describe how `renderSize` works with `latencyHint`
* Fix the enumeration description table
* Move `ScriptProcessorNode` into issues section

* Use description list for the new dictionary member

Use an HTML description list to describe the dictionary members so it looks more like the spec.

* More markup changes and add security/privacy section

* Use more description lists
* Clean up some markup in HTML parts
* Add a security/privacy section since "hardware" exposes
  finger-printable information about the user's hardware.

* Clean up HTML and fix up link to issues section

* Move security section up

* Update description of "hardware" enum

* Some cleanups

For OfflineAudioContextOptions.renderSizeHint, just say it's the same
as for AudioContextOptions, except "hardware" is the same as
"default'.

Allow more flexibility for the interaction between `latencyHint` and
`renderSize`.  I think Firefox always returns 0 for `latencyHint`
because it doesn't double-buffer.  Chrome double-buffers so the
`latencyHint` isn't zero.

Mark the example as being non-normative.

* Clean up markup

Fix typo in markup.

Slight clarification that `latencyHint` may but is not required to increase above the `renderSize`.

* Add email address

* Fix some typos and add example for 2048 frames

* Address review comments

Update according to comments from @svgeesus

* Add impl note for DelayNode

Add a section about how this affects a DelayNode.

Co-authored-by: Raymond Toy <rtoy@chromium.org>
Co-authored-by: skratchdot <git@skratchdot.com>