Fix #2361: Use new Web IDL buffer primitives (#2363)
* Use new Web IDL buffer primitives

See https://github.com/heycam/webidl/pull/987.

* Fix #2361: Use new Web IDL buffer primitives

Updates spec to use new Web IDL buffer primitives.  The changes are
marked as candidate corrections, but we use several correction
amendments to make it easy to see what the new and old text will be.
Using just one makes it super-hard to figure out what controls the
text, and the text sometimes doesn't always clearly show what's
deleted and what's inserted because the links are colored correctly,
and struck-through text doesn't strike through links.  (A bikeshed
issue?)

* Remove link that no longer exists

* Update changelog

* Add script to create list of corrections for a change

* Add comments

* Change text slightly

Change "write" to "allow writing into" so reflects better what
getChannelData allows.

* Add links to the previous and next change in the amendment box.

* Inline the JS file

* Fix up script and add event listener

Fix up the script so bikeshed doesn't mangle it.

Add an onload event listener so that when the doc is loaded we can run
ListAmendments to add a list of the related changes for the change
log.

* Fix typo

* Inline JS into index.bs

Adjust so bikeshed doesn't mangle it.

Move the listener code into the same script segment so it's easy to
find.

* Add some comments

* Refine markup

* The changes should should have the "proposed" class too.
* Update text to say "Proposed", as shown in
  https://fantasai.inkedblade.net/style/design/w3c-restyle/2020/readme#proposed-changes
* Update buttons to say "Previous Change" and "Next Change" to make it
  clearer what the buttons do.

* Refine amendment boxes

* Limit the scope of each change to a paragraph or item in an algorithm.
* Move the changed section to be in the `<div>` for the amendment box
  to it's easier to see where the changes could be.

* Fix bad numbering.

The id had the correct value, but the link for the issue used the
wrong change number (should be 3 instead of 2).

* Remove amendents.js.

Not needed because we've basically inlined it all so that the preview
shows the prev/next buttons and the list of changes for the change log.

Co-authored-by: Domenic Denicola <d@domenic.me>
Co-authored-by: Raymond Toy <rtoy@chromium.org>