Add "extract" class to old AudioContext IDL definition
The `data-no-idl` attribute that was added in #2462 tells Bikeshed not to
process the current `AudioContext` IDL block. This is enough to work around
duplicate definitions that would otherwise occur.

However, even though it is not processed, the IDL block still gets added to the
final IDL index, which thus contains both the current and future `AudioContext`
interface definitions:
https://webaudio.github.io/web-audio-api/#idl-index

The `extract` class tells Bikeshed not to include the IDL block in the final
IDL index.