Adjust del part of AudioWorklet to avoid duplication in IDL index (#2511)
Editing tools are still not very good at handling proposed changes constructs in
specifications. The `<del>` part of the `AudioWorklet` definition is incorrectly
linkified and duplicated in the final IDL index.

This update adds the `extract` class and the `data-no-idl` attributes to this
definition as a workaround, as was done for `AudioContext` and
`AudioWorkletProcessorConstructor`.